Refactor `RatingBlock`-related code
This commit moves the following classes and methods into new files,
separating and encapsulating them as appropriate:

  *  RatingBlock
  *  RatingBlockBuilder
  *  UncompressedInBlock
  *  UncompressedInBlockBuilder
  *  KeyWrapper
  *  UncompressedInBlockSort
  *  LocalIndexEncoder
  *  partitionRatings
  *  makeBlocks

In the course of this refactoring we create a new class, `RatingBlocks`,
to hold the user/item in/out block data and associated logic.
