+++ mllib/src/main/scala/org/apache/spark/ml/recommendation/UncompressedInBlock.scala
@@ -1,0 +1,230 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark.ml.recommendation
+
+import scala.collection.mutable.ArrayBuilder
+import scala.reflect.ClassTag
+
+import org.apache.spark.internal.Logging
+import org.apache.spark.util.Utils
+import org.apache.spark.util.collection.{SortDataFormat, Sorter}
+
+/**
+ * A block of (srcId, dstEncodedIndex, rating) tuples stored in primitive arrays.
+ */
+private[recommendation] class UncompressedInBlock[@specialized(Int, Long) ID: ClassTag](
+    val srcIds: Array[ID],
+    val dstEncodedIndices: Array[Int],
+    val ratings: Array[Float])(
+    implicit ord: Ordering[ID]) extends Logging {
+
+  import UncompressedInBlock.{KeyWrapper, UncompressedInBlockSort}
+
+  /** Size the of block. */
+  def length: Int = srcIds.length
+
+  /**
+   * Compresses the block into an [[InBlock]]. The algorithm is the same as converting a
+   * sparse matrix from coordinate list (COO) format into compressed sparse column (CSC) format.
+   * Sorting is done using Spark's built-in Timsort to avoid generating too many objects.
+   */
+  def compress(): InBlock[ID] = {
+    val sz = length
+    assert(sz > 0, "Empty in-link block should not exist.")
+    sort()
+    val uniqueSrcIdsBuilder = ArrayBuilder.make[ID]
+    val dstCountsBuilder = ArrayBuilder.make[Int]
+    var preSrcId = srcIds(0)
+    uniqueSrcIdsBuilder += preSrcId
+    var curCount = 1
+    var i = 1
+    while (i < sz) {
+      val srcId = srcIds(i)
+      if (srcId != preSrcId) {
+        uniqueSrcIdsBuilder += srcId
+        dstCountsBuilder += curCount
+        preSrcId = srcId
+        curCount = 0
+      }
+      curCount += 1
+      i += 1
+    }
+    dstCountsBuilder += curCount
+    val uniqueSrcIds = uniqueSrcIdsBuilder.result()
+    val numUniqueSrdIds = uniqueSrcIds.length
+    val dstCounts = dstCountsBuilder.result()
+    val dstPtrs = new Array[Int](numUniqueSrdIds + 1)
+    var sum = 0
+    i = 0
+    while (i < numUniqueSrdIds) {
+      sum += dstCounts(i)
+      i += 1
+      dstPtrs(i) = sum
+    }
+    InBlock(uniqueSrcIds, dstPtrs, dstEncodedIndices, ratings)
+  }
+
+  private def sort(): Unit = {
+    val sz = length
+    // Since there might be interleaved log messages, we insert a unique id for easy pairing.
+    val sortId = Utils.random.nextInt()
+    logDebug(s"Start sorting an uncompressed in-block of size $sz. (sortId = $sortId)")
+    val start = System.nanoTime()
+    val sorter = new Sorter(new UncompressedInBlockSort[ID])
+    sorter.sort(this, 0, length, Ordering[KeyWrapper[ID]])
+    val duration = (System.nanoTime() - start) / 1e9
+    logDebug(s"Sorting took $duration seconds. (sortId = $sortId)")
+  }
+}
+
+private[recommendation] object UncompressedInBlock {
+
+  /**
+   * Builder for uncompressed in-blocks of (srcId, dstEncodedIndex, rating) tuples.
+   *
+   * @param encoder encoder for dst indices
+   */
+  private[recommendation] class Builder[@specialized(Int, Long) ID: ClassTag](
+      encoder: LocalIndexEncoder)(
+      implicit ord: Ordering[ID]) {
+
+    private val srcIds = ArrayBuilder.make[ID]
+    private val dstEncodedIndices = ArrayBuilder.make[Int]
+    private val ratings = ArrayBuilder.make[Float]
+
+    /**
+     * Adds a dst block of (srcId, dstLocalIndex, rating) tuples.
+     *
+     * @param dstBlockId dst block ID
+     * @param srcIds original src IDs
+     * @param dstLocalIndices dst local indices
+     * @param ratings ratings
+     */
+    def add(
+        dstBlockId: Int,
+        srcIds: Array[ID],
+        dstLocalIndices: Array[Int],
+        ratings: Array[Float]): this.type = {
+      val sz = srcIds.length
+      require(dstLocalIndices.length == sz)
+      require(ratings.length == sz)
+      this.srcIds ++= srcIds
+      this.ratings ++= ratings
+      var j = 0
+      while (j < sz) {
+        this.dstEncodedIndices += encoder.encode(dstBlockId, dstLocalIndices(j))
+        j += 1
+      }
+      this
+    }
+
+    /** Builds a [[UncompressedInBlock]]. */
+    def build(): UncompressedInBlock[ID] = {
+      new UncompressedInBlock(srcIds.result(), dstEncodedIndices.result(), ratings.result())
+    }
+  }
+
+  /**
+   * A wrapper that holds a primitive key.
+   *
+   * @see [[UncompressedInBlockSort]]
+   */
+  private class KeyWrapper[@specialized(Int, Long) ID: ClassTag](
+      implicit ord: Ordering[ID]) extends Ordered[KeyWrapper[ID]] {
+
+    var key: ID = _
+
+    override def compare(that: KeyWrapper[ID]): Int = {
+      ord.compare(key, that.key)
+    }
+
+    def setKey(key: ID): this.type = {
+      this.key = key
+      this
+    }
+  }
+
+  /**
+   * [[SortDataFormat]] of [[UncompressedInBlock]] used by [[Sorter]].
+   */
+  private class UncompressedInBlockSort[@specialized(Int, Long) ID: ClassTag](
+      implicit ord: Ordering[ID])
+    extends SortDataFormat[KeyWrapper[ID], UncompressedInBlock[ID]] {
+
+    override def newKey(): KeyWrapper[ID] = new KeyWrapper()
+
+    override def getKey(
+        data: UncompressedInBlock[ID],
+        pos: Int,
+        reuse: KeyWrapper[ID]): KeyWrapper[ID] = {
+      if (reuse == null) {
+        new KeyWrapper().setKey(data.srcIds(pos))
+      } else {
+        reuse.setKey(data.srcIds(pos))
+      }
+    }
+
+    override def getKey(
+        data: UncompressedInBlock[ID],
+        pos: Int): KeyWrapper[ID] = {
+      getKey(data, pos, null)
+    }
+
+    private def swapElements[@specialized(Int, Float) T](
+        data: Array[T],
+        pos0: Int,
+        pos1: Int): Unit = {
+      val tmp = data(pos0)
+      data(pos0) = data(pos1)
+      data(pos1) = tmp
+    }
+
+    override def swap(data: UncompressedInBlock[ID], pos0: Int, pos1: Int): Unit = {
+      swapElements(data.srcIds, pos0, pos1)
+      swapElements(data.dstEncodedIndices, pos0, pos1)
+      swapElements(data.ratings, pos0, pos1)
+    }
+
+    override def copyRange(
+        src: UncompressedInBlock[ID],
+        srcPos: Int,
+        dst: UncompressedInBlock[ID],
+        dstPos: Int,
+        length: Int): Unit = {
+      System.arraycopy(src.srcIds, srcPos, dst.srcIds, dstPos, length)
+      System.arraycopy(src.dstEncodedIndices, srcPos, dst.dstEncodedIndices, dstPos, length)
+      System.arraycopy(src.ratings, srcPos, dst.ratings, dstPos, length)
+    }
+
+    override def allocate(length: Int): UncompressedInBlock[ID] = {
+      new UncompressedInBlock(
+        new Array[ID](length), new Array[Int](length), new Array[Float](length))
+    }
+
+    override def copyElement(
+        src: UncompressedInBlock[ID],
+        srcPos: Int,
+        dst: UncompressedInBlock[ID],
+        dstPos: Int): Unit = {
+      dst.srcIds(dstPos) = src.srcIds(srcPos)
+      dst.dstEncodedIndices(dstPos) = src.dstEncodedIndices(srcPos)
+      dst.ratings(dstPos) = src.ratings(srcPos)
+    }
+  }
+
+}
